Moved test classes to banking package.
Moved all of the test classes to the banking package to allow access to any package private fields or methods. Fixed a couple other small errors that were throwing flags in IDE.
